% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsvEst.R
\name{bsvSimu}
\alias{bsvSimu}
\title{A Cat Function}
\usage{
bsvSimu(mu1, phi1, sigma1, length1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
