\name{Stocastic Volatility Package}
\alias{SV-package}
\alias{SV}
\docType{package}
\title{
<<<<<<< HEAD
Mixture sampler for stochastic volatility models
}
\description{
=======
\packageTitle{SV}
Mixture sampler for stochastic volatility models
}
\description{
\packageDescription{SV}
>>>>>>> 38b6ed05deb6eacf7e01da9353876c2da72eacbc
It will implement Markov chain Monte Carlo simulation for stochastic volatility models using the mixture sampler by Omori, Chib, Shephard and Nakajima (2007).
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SV}
\packageIndices{SV}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
<<<<<<< HEAD
Ryuji Hashimoto and Yasuhiro Omori. 
Maintainer: Yasuhiro Omori<omori@e.u-tokyo.ac.jp>
=======
\packageAuthor{SV}

Maintainer: \packageMaintainer{SV} Yasuhiro Omori
>>>>>>> 38b6ed05deb6eacf7e01da9353876c2da72eacbc
}
\references{
Omori, Y., Chib, S., Shephard, N., & Nakajima, J. (2007). Stochastic volatility with leverage: Fast and efficient likelihood inference. Journal of Econometrics, 140(2), 425-449.
}
<<<<<<< HEAD
\keyword{ Markov chain Monte Carlo. Mixture Sampler. Stochastic Volatility. }
=======
\keyword{ Markov chain Monte Carlo, Mixture Sampler, Stochastic Volatility }
>>>>>>> 38b6ed05deb6eacf7e01da9353876c2da72eacbc
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
svmcmc(data, TRUE)
}
