\name{SV-package}
\alias{SV-package}
\alias{SV}
\docType{package}
\title{
\packageTitle{SV}
}
\description{
\packageDescription{SV}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SV}
\packageIndices{SV}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{SV}

Maintainer: \packageMaintainer{SV}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
